requirements.txt
aiofiles==23.2.1
aiogram==3.22.0
aiogram_calendar==0.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.19.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.2
jiter==0.10.0
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
numpy==2.0.2
openai==1.102.0
packaging==25.0
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.7
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.24.0
yarl==1.20.1
